Add files via upload
This code defines a 2D maze as a list of lists and uses depth-first search (DFS) and breadth-first search (BFS) algorithms to find the path from the starting point (0,0) to the goal point (4,3) in the maze.

The BFS and DFS algorithms are defined as functions and take the maze as an argument. The DFS function uses a stack to explore neighboring cells and the BFS function uses a queue. Both functions keep track of the visited cells and the path from the start to the current cell.

The code also defines a get_neighbors function that takes a cell and the maze as arguments and returns a list of neighboring cells that can be visited.

Finally, the code calls both the BFS and DFS functions and prints out the cells visited during the algorithm. The path from the start to the goal cell is returned as a list of tuples representing the cells in the path.